Move StudioMesh to Libs/Analyze.  Needs to be renamed and/or merged into Mesh class.
